switch from proxy to redeploy
