+++ stack
@@ -198,16 +198,17 @@ function boot() {
 ############################################################################
 ############################################################################
 
-function upgrade-storage() {
+function redeploy-storage() {
   source .env
   eval $(./stack print-local-dev-env)
+  export WEB3_PRIVATE_KEY=$ADMIN_PRIVATE_KEY
   #export WEB3_RPC_URL=$WEB3_RPC_URL
   #export WEB3_CHAIN_ID=$WEB3_CHAIN_ID
   export NETWORK=$NETWORK
   (
     set -euo pipefail
     cd hardhat
-    npx hardhat run scripts/upgrade-storage.ts --network "$NETWORK"
+    npx hardhat run scripts/redeploy-storage.ts --network "$NETWORK"
   )
 }
